// lib/schemas/federation/ClientConfig.ts
/*
 * Copyright (C) 2014-2024 Authlete, Inc.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import { z } from 'zod';

/**
 * Schema for client configuration used in federation flows.
 */
export const clientConfigSchema = z.object({
  /**
   * The client identifier.
   */
  clientId: z.string().min(1),

  /**
   * The client secret.
   */
  clientSecret: z.string().min(1),

  /**
   * The redirect URI where the authorization response will be sent.
   */
  redirectUri: z.string().url(),

  /**
   * The algorithm used to sign the ID token.
   */
  idTokenSignedResponseAlg: z.string().nullish(),
});

/**
 * Type definition for ClientConfig.
 */
export type ClientConfig = z.infer<typeof clientConfigSchema>;